Reorganized signup
